requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
requests>=2.31.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
plyer>=2.1.0
pydub>=0.25.1
pydantic>=2.4
python-dotenv>=1.0
